refactor: normalize folder names!
