+++ app/components/DynamicSvg/DynamicSvg.js
