+++ app/components/ResultFeedback/ResultFeedback.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useContext } from 'react'
-import { AppContext } from '../../Context/AppContext'
+import { AppContext } from '../../context/AppContext'
 import Spinner from '../Loader/Spinner'
 
 const ResultFeedback = () => {
